+++ src/lib/rmc-picker/es/PopupMixin.js
@@ -1,49 +1,18 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-    value: true
-});
-
-var _extends2 = require('babel-runtime/helpers/extends');
-
-var _extends3 = _interopRequireDefault(_extends2);
-
-var _defineProperty2 = require('babel-runtime/helpers/defineProperty');
-
-var _defineProperty3 = _interopRequireDefault(_defineProperty2);
-
-var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
-
-var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
-
-var _createClass2 = require('babel-runtime/helpers/createClass');
-
-var _createClass3 = _interopRequireDefault(_createClass2);
-
-var _possibleConstructorReturn2 = require('babel-runtime/helpers/possibleConstructorReturn');
-
-var _possibleConstructorReturn3 = _interopRequireDefault(_possibleConstructorReturn2);
-
-var _inherits2 = require('babel-runtime/helpers/inherits');
-
-var _inherits3 = _interopRequireDefault(_inherits2);
-
-exports['default'] = PopupMixin;
-
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-function PopupMixin(getModal, platformProps) {
+import _extends from 'babel-runtime/helpers/extends';
+import _defineProperty from 'babel-runtime/helpers/defineProperty';
+import _classCallCheck from 'babel-runtime/helpers/classCallCheck';
+import _createClass from 'babel-runtime/helpers/createClass';
+import _possibleConstructorReturn from 'babel-runtime/helpers/possibleConstructorReturn';
+import _inherits from 'babel-runtime/helpers/inherits';
+import React from 'react';
+export default function PopupMixin(getModal, platformProps) {
     return _a = function (_React$Component) {
-        (0, _inherits3['default'])(_a, _React$Component);
+        _inherits(_a, _React$Component);
 
         function _a(props) {
-            (0, _classCallCheck3['default'])(this, _a);
+            _classCallCheck(this, _a);
 
-            var _this = (0, _possibleConstructorReturn3['default'])(this, (_a.__proto__ || Object.getPrototypeOf(_a)).call(this, props));
+            var _this = _possibleConstructorReturn(this, (_a.__proto__ || Object.getPrototypeOf(_a)).call(this, props));
 
             _this.onPickerChange = function (pickerValue) {
                 if (_this.state.pickerValue !== pickerValue) {
@@ -83,7 +52,7 @@ function PopupMixin(getModal, platformProps) {
                     if (pickerValue === null) {
                         pickerValue = _this.props.value;
                     }
-                    return _react2['default'].cloneElement(_this.props.picker, (_React$cloneElement = {}, (0, _defineProperty3['default'])(_React$cloneElement, _this.props.pickerValueProp, pickerValue), (0, _defineProperty3['default'])(_React$cloneElement, _this.props.pickerValueChangeProp, _this.onPickerChange), (0, _defineProperty3['default'])(_React$cloneElement, 'ref', _this.saveRef), _React$cloneElement));
+                    return React.cloneElement(_this.props.picker, (_React$cloneElement = {}, _defineProperty(_React$cloneElement, _this.props.pickerValueProp, pickerValue), _defineProperty(_React$cloneElement, _this.props.pickerValueChangeProp, _this.onPickerChange), _defineProperty(_React$cloneElement, 'ref', _this.saveRef), _React$cloneElement));
                 } else {
                     return _this.props.content;
                 }
@@ -102,7 +71,7 @@ function PopupMixin(getModal, platformProps) {
             return _this;
         }
 
-        (0, _createClass3['default'])(_a, [{
+        _createClass(_a, [{
             key: 'componentWillReceiveProps',
             value: function componentWillReceiveProps(nextProps) {
                 if ('value' in nextProps) {
@@ -158,10 +127,10 @@ function PopupMixin(getModal, platformProps) {
                 if (!disabled) {
                     newChildProps[props.triggerType] = this.onTriggerClick;
                 }
-                return _react2['default'].createElement(
+                return React.createElement(
                     WrapComponent,
                     { style: props.wrapStyle },
-                    _react2['default'].cloneElement(child, newChildProps),
+                    React.cloneElement(child, newChildProps),
                     getModal(props, this.state.visible, {
                         getContent: this.getContent,
                         onOk: this.onOk,
@@ -176,12 +145,12 @@ function PopupMixin(getModal, platformProps) {
                 return this.getRender();
             }
         }]);
+
         return _a;
-    }(_react2['default'].Component), _a.defaultProps = (0, _extends3['default'])({
+    }(React.Component), _a.defaultProps = _extends({
         onVisibleChange: function onVisibleChange(_) {},
         okText: 'Ok', dismissText: 'Dismiss', title: '', onOk: function onOk(_) {},
         onDismiss: function onDismiss() {}
     }, platformProps), _a;
     var _a;
 }
-module.exports = exports['default'];