+++ src/lib/rmc-picker/lib/MultiPicker.native.js
@@ -4,17 +4,11 @@ Object.defineProperty(exports, "__esModule", {
     value: true
 });
 
-var _extends2 = require('babel-runtime/helpers/extends');
-
-var _extends3 = _interopRequireDefault(_extends2);
-
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
-var _classnames = require('classnames');
-
-var _classnames2 = _interopRequireDefault(_classnames);
+var _reactNative = require('react-native');
 
 var _MultiPickerMixin = require('./MultiPickerMixin');
 
@@ -23,10 +17,7 @@ var _MultiPickerMixin2 = _interopRequireDefault(_MultiPickerMixin);
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
 var MultiPicker = function MultiPicker(props) {
-    var prefixCls = props.prefixCls,
-        className = props.className,
-        rootNativeProps = props.rootNativeProps,
-        children = props.children,
+    var children = props.children,
         style = props.style;
 
     var selectedValue = props.getValue();
@@ -39,19 +30,12 @@ var MultiPicker = function MultiPicker(props) {
                 }
 
                 return props.onValueChange.apply(props, [i].concat(args));
-            },
-            onScrollChange: props.onScrollChange && function () {
-                for (var _len2 = arguments.length, args = Array(_len2), _key2 = 0; _key2 < _len2; _key2++) {
-                    args[_key2] = arguments[_key2];
-                }
-
-                return props.onScrollChange.apply(props, [i].concat(args));
             }
         });
     });
     return _react2['default'].createElement(
-        'div',
-        (0, _extends3['default'])({}, rootNativeProps, { style: style, className: (0, _classnames2['default'])(className, prefixCls) }),
+        _reactNative.View,
+        { style: style },
         colElements
     );
 };