+++ src/lib/rmc-picker/lib/NativePicker.js
@@ -9,10 +9,10 @@ var _NativePicker = require('./NativePicker.ios');
 Object.defineProperty(exports, 'default', {
   enumerable: true,
   get: function get() {
-    return _interopRequireDefault(_NativePicker)["default"];
+    return _interopRequireDefault(_NativePicker)['default'];
   }
 });
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
 module.exports = exports['default'];