+++ src/lib/rmc-picker/lib/index.js
@@ -9,26 +9,10 @@ var _Picker = require('./Picker');
 Object.defineProperty(exports, 'default', {
   enumerable: true,
   get: function get() {
-    return _interopRequireDefault(_Picker)["default"];
+    return _interopRequireDefault(_Picker)['default'];
   }
 });
 
-var _Popup = require('./Popup');
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
-Object.defineProperty(exports, 'Popup', {
-  enumerable: true,
-  get: function get() {
-    return _interopRequireDefault(_Popup)["default"];
-  }
-});
-
-var _PopupStyles = require('./PopupStyles');
-
-Object.defineProperty(exports, 'PopupStyles', {
-  enumerable: true,
-  get: function get() {
-    return _interopRequireDefault(_PopupStyles)["default"];
-  }
-});
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
+module.exports = exports['default'];